Refactor some classes in utilties file to separate files
