+++ Core/CRC16.cs
@@ -1,0 +1,62 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Substructio.Core
+{
+    public class CRC16
+    {
+        const ushort polynomial = 0xA001;
+        ushort[] table = new ushort[256];
+
+        private static CRC16 _instance;
+
+        public int ComputeChecksum(byte[] bytes)
+        {
+            ushort crc = 0;
+            for (int i = 0; i < bytes.Length; ++i)
+            {
+                byte index = (byte)(crc ^ bytes[i]);
+                crc = (ushort)((crc >> 8) ^ table[index]);
+            }
+            return crc;
+        }
+
+        public byte[] ComputeChecksumBytes(byte[] bytes)
+        {
+            int crc = ComputeChecksum(bytes);
+            return BitConverter.GetBytes(crc);
+        }
+
+        public CRC16()
+        {
+            ushort value;
+            ushort temp;
+            for (ushort i = 0; i < table.Length; ++i)
+            {
+                value = 0;
+                temp = i;
+                for (byte j = 0; j < 8; ++j)
+                {
+                    if (((value ^ temp) & 0x0001) != 0)
+                    {
+                        value = (ushort)((value >> 1) ^ polynomial);
+                    }
+                    else
+                    {
+                        value >>= 1;
+                    }
+                    temp >>= 1;
+                }
+                table[i] = value;
+            }
+        }
+
+        public static CRC16 Instance()
+        {
+            return _instance ?? (_instance = new CRC16());
+        }
+    }
+}
